Delete superfluous helper file
